requirements.txt
langchain>=0.1.10
langchain-community>=0.0.25
ollama>=0.1.7
streamlit>=1.31.1
streamlit_chat>=0.1.1
pypdf>=4.0.2
fastembed>=v0.2.2
chromadb>=0.4.24
python-dotenv>=1.0.1